ISSUE 227: add `parser_settings.debug` flag and enable `csv2` reader to optionally inject debug (line) info into record IDR. (#228)
Decided to make the debug flag a global setting in `parser_settings` and leave its type to be an int (>= 0) for
future flexibility. Could've done with a string/enum to make it more defined and strict, but given it is an adv
setting and current usage is so scarce thus leaving it flexible until further requirements arise.

For `csv2` reader, if `parser_settings.debug` is 0 or omitted, which is the vast vast majority of existing and
future `csv2` schemas, no behavior changes; if `parser_settings.debug` isn't 0, then a `__debug` node will
be added to the record IDR structure, underneath which, currently only `line_num` debug info will be added.

This design is flexible for all future adoptions in all other file format readers, yet has zero impact on any existing
schemas.